api/requirements.txt
google-generativeai==0.3.2
python-dateutil==2.8.2
